+++ infrastructure/server/oidc_clients_spar.tf
@@ -20,7 +20,7 @@ resource "aws_cognito_user_pool_client" "dev_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_dev"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  read_attributes                               = var.minimum_oidc_attribute_list
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.dev_idir_oidc_provider.provider_name}",
@@ -34,7 +34,7 @@ resource "aws_cognito_user_pool_client" "dev_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  write_attributes = var.minimum_oidc_attribute_list
 }
 
 resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
@@ -66,7 +66,7 @@ resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_test"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  read_attributes                               = var.minimum_oidc_attribute_list
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.test_idir_oidc_provider.provider_name}",
@@ -80,7 +80,7 @@ resource "aws_cognito_user_pool_client" "test_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  write_attributes = var.minimum_oidc_attribute_list
 }
 
 resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
@@ -104,7 +104,7 @@ resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
   id_token_validity                             = "60"
   name                                          = "spar_prod"
   prevent_user_existence_errors                 = "ENABLED"
-  read_attributes                               = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  read_attributes                               = var.minimum_oidc_attribute_list
   refresh_token_validity                        = "24"
   supported_identity_providers                  = [
     "${aws_cognito_identity_provider.prod_idir_oidc_provider.provider_name}",
@@ -118,5 +118,5 @@ resource "aws_cognito_user_pool_client" "prod_spar_oidc_client" {
   }
 
   user_pool_id     = aws_cognito_user_pool.fam_user_pool.id
-  write_attributes = "${concat(var.minimum_oidc_attribute_list, ["custom:idp_display_name", "email"])}"
+  write_attributes = var.minimum_oidc_attribute_list
 }