+++ infrastructure/server/variables_provided.tf
@@ -259,10 +259,21 @@ variable "fam_console_idp_name_bceid" {
   type = string
 }
 
+/*
+  Note!: Required for all OIDC clients.
+  All attributes of the user gets stored/updated in FAM_USER table must be part of the
+  minimum attribute set used by all applications. This is to avoid attributes being
+  overriden (when the same user can login through different app clients) on Cognito User
+  Pool for that user and avoid attributes being overriden into FAM_USER table.
+*/
 variable "minimum_oidc_attribute_list" {
   description = "Required fields for FAM clients to be able to read and write"
   type        = list(string)
-  default     = ["custom:idp_name", "custom:idp_user_id", "custom:idp_username"]
+  default     = [
+    "custom:idp_name", "custom:idp_user_id", "custom:idp_username",
+    "custom:idp_business_id", "custom:idp_display_name", "email",
+    "given_name", "family_name"
+  ]
 }
 
 variable "maximum_oidc_attribute_read_list" {